moves execution model to cli package
